Add files via upload
Updated flash cards to complete the entire dictionary in random order (rather than choosing random words from the dictionary) keeping track of incorrect answers. Once complete with the dictionary, goes through the incorrect answers, keeping track if incorrect answers, so on and so forth until the original list is loaded again in random order (if all of the answers are done correctly).